feat(newOnboarding): changes Title style
