+++ src/components/organisms/OnBoarding/components/Content.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const Content = styled.div`
+  width: 100%;
+  max-width: 768px;
+`;
+
+export default Content;
