+++ src/components/organisms/OnBoarding/components/Landing.tsx
@@ -3,5 +3,8 @@ import styled from 'styled-components';
 export default styled('section')`
   height: 100vh;
   text-align: center;
-  font-family: 'Lato', sans-serif;
+
+  h1 {
+    font-style: italic;
+  }
 `;
